➕ added zustand and implemented it
