+++ apps/mandelbrot-react/package.json
@@ -10,8 +10,10 @@
     "node-sass": "^4.14.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-reflex": "^3.1.0",
     "react-resizable": "^1.11.0",
-    "react-scripts": "3.4.3"
+    "react-scripts": "3.4.3",
+    "zustand": "^3.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
