+++ apps/mandelbrot-react/src/components/Options.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import Renderer from './Renderer'
+
+const Options = () => {
+	return (
+		<div>
+			<Renderer />
+		</div>
+	)
+}
+
+export default Options
